requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
pyotp
qrcode
pillow